kernel etc + troep

+++ BlackTracer2/BlackTracer/Star.h
@@ -3,13 +3,13 @@
 class Star
 {
 public:
-	double phi;
-	double theta;
-	double magnitude;
+	float phi;
+	float theta;
+	float magnitude;
 	Vec3b color;
-	Point2d posInPix;
+	Point2f posInPix;
 
-	Star(double p, double t, double _magnitude, Vec3b _color) {
+	Star(float p, float t, float _magnitude, Vec3b _color) {
 		phi = p;
 		theta = t;
 		magnitude = _magnitude;
@@ -33,9 +33,9 @@ public:
 			// Compute individual hash values for first, second and third
 			// http://stackoverflow.com/a/1646913/126995
 			size_t res = 17;
-			res = res * 31 + hash<double>()(star.theta);
-			res = res * 31 + hash<double>()(star.phi);
-			res = res * 31 + hash<double>()(star.magnitude);
+			res = res * 31 + hash<float>()(star.theta);
+			res = res * 31 + hash<float>()(star.phi);
+			res = res * 31 + hash<float>()(star.magnitude);
 			return res;
 		}
 	};
